Added Resources, Ran Grammarly
As proposed by @gitr0n1n, added the "Ethereum Classic Quickstart" to the basics page where it fits slightly better than to the "News" page. Also, I realized that the Solidity Smart Contract section is a bit sparse, so I added a "Solidity Crash Course" (free YT video+blog lessons, no paywall). As I was reading over the article, I also found some grammar mistakes and clunky phrases which I took the liberty of correcting to the best of my knowledge. And I linked to Wikipedia for the phrase "Turing-Completeness".